+++ src/shader/shaderStructs.js
@@ -183,7 +183,7 @@ export const shaderMaterialStructs = /* glsl */ `
 		m.matte = bool( s13.r );
 		m.castShadow = ! bool( s13.g );
 
-		uint firstTextureTransformIdx = i + 12u;
+		uint firstTextureTransformIdx = i + 14u;
 
 		m.mapTransform = m.map == - 1 ? mat3( 0 ) : readTextureTransform( tex, firstTextureTransformIdx );
 		m.metalnessMapTransform = m.metalnessMap == - 1 ? mat3( 0 ) : readTextureTransform( tex, firstTextureTransformIdx + 2u );
@@ -208,6 +208,18 @@ export const shaderMaterialStructs = /* glsl */ `
 `;
 
 export const shaderLightStruct = /* glsl */ `
+
+	#define RECT_AREA_LIGHT_TYPE 0
+	#define CIRC_AREA_LIGHT_TYPE 1
+	#define SPOT_LIGHT_TYPE 2
+
+	struct LightsInfo {
+
+		sampler2D tex;
+		uint count;
+
+	};
+
 	struct Light {
 
 		vec3 position;
@@ -220,11 +232,20 @@ export const shaderLightStruct = /* glsl */ `
 		vec3 v;
 		float area;
 
+		// spot light fields
+		float radius;
+		float near;
+		float decay;
+		float distance;
+		float coneCos;
+		float penumbraCos;
+		int iesProfile;
+
 	};
 
 	Light readLightInfo( sampler2D tex, uint index ) {
 
-		uint i = index * 4u;
+		uint i = index * 6u;
 
 		vec4 s0 = texelFetch1D( tex, i + 0u );
 		vec4 s1 = texelFetch1D( tex, i + 1u );
@@ -241,6 +262,21 @@ export const shaderLightStruct = /* glsl */ `
 		l.u = s2.rgb;
 		l.v = s3.rgb;
 		l.area = s3.a;
+
+		if ( l.type == SPOT_LIGHT_TYPE ) {
+
+			vec4 s4 = texelFetch1D( tex, i + 4u );
+			vec4 s5 = texelFetch1D( tex, i + 5u );
+			l.radius = s4.r;
+			l.near = s4.g;
+			l.decay = s4.b;
+			l.distance = s4.a;
+
+			l.coneCos = s5.r;
+			l.penumbraCos = s5.g;
+			l.iesProfile = int( round ( s5.b ) );
+
+		}
 
 		return l;
 
@@ -267,40 +303,5 @@ export const shaderLightStruct = /* glsl */ `
 		int iesProfile;
 
 	};
-
-	SpotLight readSpotLightInfo( sampler2D tex, uint index ) {
-
-		uint i = index * 6u;
-
-		vec4 s0 = texelFetch1D( tex, i + 0u );
-		vec4 s1 = texelFetch1D( tex, i + 1u );
-		vec4 s2 = texelFetch1D( tex, i + 2u );
-		vec4 s3 = texelFetch1D( tex, i + 3u );
-		vec4 s4 = texelFetch1D( tex, i + 4u );
-		vec4 s5 = texelFetch1D( tex, i + 5u );
-
-		SpotLight sl;
-		sl.position = s0.rgb;
-		sl.type = int( round( s0.a ) );
-
-		sl.color = s1.rgb;
-		sl.intensity = s1.a;
-
-		sl.u = s2.rgb;
-		sl.v = s3.rgb;
-		sl.area = s3.a;
-
-		sl.radius = s4.r;
-		sl.near = s4.g;
-		sl.decay = s4.b;
-		sl.distance = s4.a;
-
-		sl.coneCos = s5.r;
-		sl.penumbraCos = s5.g;
-		sl.iesProfile = int( round ( s5.b ) );
-
-		return sl;
-
-	}
 
 `;